v0.5.1: FAT open(), read() and close()
New:

- kerrno: errno implementation for the Garn kernel.

- FAT: added open(), read() and close() routines.

- FAL: Added the option to select the seek origin of kfseek via the "whence" argument.

Fixes:

- kheap: fixed a bug where allocated blocks that aren't next to eachother would get merged and possibly overwrite the data in between them.

- Console Demo: fixed a bug that caused UB when registering a new command in a previously allocated kheap block.

+++ kernel/sys/fal/filesys.h
@@ -31,11 +31,11 @@ typedef struct _filesys {
     size_t size;
 
     struct _file* (*open)(struct _filesys* self, char* path, uint8_t access);
-    uint8_t (*read)(struct _filesys* self, struct _file* file, size_t size, void* buf);
-    uint8_t (*write)(struct _filesys* self, struct _file* file, size_t size, void* buf);
-    uint8_t (*close)(struct _filesys* self, struct _file* file);
-    uint8_t (*mkdir)(struct _filesys* self, char* path);
-    uint8_t (*rmdir)(struct _filesys* self, char* path);
+    int (*read)(struct _filesys* self, struct _file* file, size_t size, void* buf);
+    int (*write)(struct _filesys* self, struct _file* file, size_t size, void* buf);
+    int (*close)(struct _filesys* self, struct _file* file);
+    int (*mkdir)(struct _filesys* self, char* path);
+    int (*rmdir)(struct _filesys* self, char* path);
 
     //should not be touched by drivers
     bool _valid;
